Move mapping code to a seperate module
