+++ src/ck-core/cklocation.ci
@@ -1,5 +1,6 @@
 module CkLocation {
   extern module CkMarshall;
+  extern module CkArrayMap;
 
   message CkArrayElementMigrateMessage {
     char packData[];
@@ -20,57 +21,5 @@ module CkLocation {
 
     //Load balancing:
     entry void dummyAtSync(void);
-  };
-  
-  // Array Map object support:
-  group [migratable] CkArrayMap {
-  };
-
-  group [migratable] RRMap : CkArrayMap {
-    entry RRMap(void);
-  };
-
-  readonly CkGroupID _defaultArrayMapID;
-  readonly CkGroupID _fastArrayMapID;
-
-  group [migratable] DefaultArrayMap : RRMap {
-    entry DefaultArrayMap(void);
-  };
-
-  group [migratable] FastArrayMap : DefaultArrayMap {
-    entry FastArrayMap(void);
-  };
-
-  group [migratable] HilbertArrayMap : DefaultArrayMap {
-    entry HilbertArrayMap(void);
-  };
-
-  group [migratable] ReadFileMap : DefaultArrayMap {
-    entry ReadFileMap(void);
-  };
-
-  group [migratable] BlockMap : DefaultArrayMap {
-    entry BlockMap(void);
-  };
-
-  group [migratable] CldMap : CkArrayMap {
-    entry CldMap(void);
-  };
-
-  group [migratable] ConfigurableRRMap : RRMap {
-    entry ConfigurableRRMap(void);
-  };
-
-  initproc void _initConfigurableRRMap(void);
-
-  initnode void _propMapInit(void);
-  initproc void _CkMigratable_initInfoInit(void);
-
-  group [migratable] PropMap : CkArrayMap {
-    entry PropMap(void);
-  };
-
-  mainchare CkMapsInit {
-    entry CkMapsInit(CkArgMsg *m);
   };
 };
